+++ EA/test/TestNetworkMapping.cpp
@@ -5,22 +5,22 @@
 
 TEST(LineNetwork, ValidNodes)
 {
-   int from_edges[2][2] = 
+   double from_edges[2][2] = 
    {
       {0, 1},
       {1, 0}
    };
-   int from_weights[2] = {1, 1};
-   int *from_matrix[2];
+   double from_weights[2] = {1, 1};
+   double *from_matrix[2];
    for (int i = 0; i < 2; i++)
    {
       from_matrix[i] = from_edges[i];
    }
  
-   Network from_network((int**)from_matrix, from_weights, 2);
+   Network from_network((double**)from_matrix, from_weights, 2);
 
 
-   int to_edges[5][5] =
+   double to_edges[5][5] =
    { 
      {0, 2, 0, 0, 0},
      {2, 0, 2, 0, 0},
@@ -29,14 +29,14 @@ TEST(LineNetwork, ValidNodes)
      {0, 0, 0, 2, 0}
    };
 
-   int to_weights[5] = {1, 1, 1, 1, 1};
-   int *to_matrix[5];
+   double to_weights[5] = {1, 1, 1, 1, 1};
+   double *to_matrix[5];
    for (int i = 0; i < 5; i++)
    {
       to_matrix[i] = to_edges[i];
    }
  
-   Network to_network((int**)to_matrix, to_weights, 5);
+   Network to_network((double**)to_matrix, to_weights, 5);
    
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 0;
@@ -47,22 +47,22 @@ TEST(LineNetwork, ValidNodes)
 
 TEST(LineNetwork, ValidNodesRepeated)
 {
-   int from_edges[2][2] = 
+   double from_edges[2][2] = 
    {
       {0, 1},
       {1, 0}
    };
-   int from_weights[2] = {1, 1};
-   int *from_matrix[2];
+   double from_weights[2] = {1, 1};
+   double *from_matrix[2];
    for (int i = 0; i < 2; i++)
    {
       from_matrix[i] = from_edges[i];
    }
  
-   Network from_network((int**)from_matrix, from_weights, 2);
+   Network from_network((double**)from_matrix, from_weights, 2);
 
 
-   int to_edges[5][5] =
+   double to_edges[5][5] =
    { 
      {0, 2, 0, 0, 0},
      {2, 0, 2, 0, 0},
@@ -71,14 +71,14 @@ TEST(LineNetwork, ValidNodesRepeated)
      {0, 0, 0, 2, 0}
    };
 
-   int to_weights[5] = {1, 1, 1, 1, 1};
-   int *to_matrix[5];
+   double to_weights[5] = {1, 1, 1, 1, 1};
+   double *to_matrix[5];
    for (int i = 0; i < 5; i++)
    {
       to_matrix[i] = to_edges[i];
    }
  
-   Network to_network((int**)to_matrix, to_weights, 5);
+   Network to_network((double**)to_matrix, to_weights, 5);
    
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 0;
@@ -92,22 +92,22 @@ TEST(LineNetwork, ValidNodesRepeated)
 
 TEST(LineNetwork, ValidLinks)
 {
-   int from_edges[2][2] = 
+   double from_edges[2][2] = 
    {
       {0, 1},
       {1, 0}
    };
-   int from_weights[2] = {1, 1};
-   int *from_matrix[2];
+   double from_weights[2] = {1, 1};
+   double *from_matrix[2];
    for (int i = 0; i < 2; i++)
    {
       from_matrix[i] = from_edges[i];
    }
  
-   Network from_network((int**)from_matrix, from_weights, 2);
+   Network from_network((double**)from_matrix, from_weights, 2);
 
 
-   int to_edges[5][5] =
+   double to_edges[5][5] =
    { 
      {0, 2, 0, 0, 0},
      {2, 0, 2, 0, 0},
@@ -116,14 +116,14 @@ TEST(LineNetwork, ValidLinks)
      {0, 0, 0, 2, 0}
    };
 
-   int to_weights[5] = {1, 1, 1, 1, 1};
-   int *to_matrix[5];
+   double to_weights[5] = {1, 1, 1, 1, 1};
+   double *to_matrix[5];
    for (int i = 0; i < 5; i++)
    {
       to_matrix[i] = to_edges[i];
    }
  
-   Network to_network((int**)to_matrix, to_weights, 5);
+   Network to_network((double**)to_matrix, to_weights, 5);
    
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 0;
@@ -134,22 +134,22 @@ TEST(LineNetwork, ValidLinks)
 
 TEST(LineNetwork, ValidLinksRepeated)
 {
-   int from_edges[2][2] = 
+   double from_edges[2][2] = 
    {
       {0, 1},
       {1, 0}
    };
-   int from_weights[2] = {1, 1};
-   int *from_matrix[2];
+   double from_weights[2] = {1, 1};
+   double *from_matrix[2];
    for (int i = 0; i < 2; i++)
    {
       from_matrix[i] = from_edges[i];
    }
  
-   Network from_network((int**)from_matrix, from_weights, 2);
+   Network from_network((double**)from_matrix, from_weights, 2);
 
 
-   int to_edges[5][5] =
+   double to_edges[5][5] =
    { 
      {0, 2, -1, -1, -1},
      {2, 0, 2, -1, -1},
@@ -158,14 +158,14 @@ TEST(LineNetwork, ValidLinksRepeated)
      {-1, -1, -1, 2, 0}
    };
 
-   int to_weights[5] = {1, 1, 1, 1, 1};
-   int *to_matrix[5];
+   double to_weights[5] = {1, 1, 1, 1, 1};
+   double *to_matrix[5];
    for (int i = 0; i < 5; i++)
    {
       to_matrix[i] = to_edges[i];
    }
  
-   Network to_network((int**)to_matrix, to_weights, 5);
+   Network to_network((double**)to_matrix, to_weights, 5);
    
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 0;
@@ -178,22 +178,22 @@ TEST(LineNetwork, ValidLinksRepeated)
 
 TEST(LineNetwork, SameNodeValidMapping)
 {
-   int from_edges[2][2] = 
+   double from_edges[2][2] = 
    {
       {0, 1},
       {1, 0}
    };
-   int from_weights[2] = {1, 1};
-   int *from_matrix[2];
+   double from_weights[2] = {1, 1};
+   double *from_matrix[2];
    for (int i = 0; i < 2; i++)
    {
       from_matrix[i] = from_edges[i];
    }
  
-   Network from_network((int**)from_matrix, from_weights, 2);
+   Network from_network((double**)from_matrix, from_weights, 2);
 
 
-   int to_edges[5][5] =
+   double to_edges[5][5] =
    { 
      {0, 2, 0, 0, 0},
      {2, 0, 2, 0, 0},
@@ -202,14 +202,14 @@ TEST(LineNetwork, SameNodeValidMapping)
      {0, 0, 0, 2, 0}
    };
 
-   int to_weights[5] = {1, 2, 1, 1, 1};
-   int *to_matrix[5];
+   double to_weights[5] = {1, 2, 1, 1, 1};
+   double *to_matrix[5];
    for (int i = 0; i < 5; i++)
    {
       to_matrix[i] = to_edges[i];
    }
  
-   Network to_network((int**)to_matrix, to_weights, 5);
+   Network to_network((double**)to_matrix, to_weights, 5);
    
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 1;
@@ -222,22 +222,22 @@ TEST(LineNetwork, SameNodeValidMapping)
 
 TEST(LineNetwork, SameNodeInvalidMapping)
 {
-   int from_edges[2][2] = 
+   double from_edges[2][2] = 
    {
       {0, 1},
       {1, 0}
    };
-   int from_weights[2] = {1, 1};
-   int *from_matrix[2];
+   double from_weights[2] = {1, 1};
+   double *from_matrix[2];
    for (int i = 0; i < 2; i++)
    {
       from_matrix[i] = from_edges[i];
    }
  
-   Network from_network((int**)from_matrix, from_weights, 2);
+   Network from_network((double**)from_matrix, from_weights, 2);
 
 
-   int to_edges[5][5] =
+   double to_edges[5][5] =
    { 
      {0, 2, 0, 0, 0},
      {2, 0, 2, 0, 0},
@@ -246,14 +246,14 @@ TEST(LineNetwork, SameNodeInvalidMapping)
      {0, 0, 0, 2, 0}
    };
 
-   int to_weights[5] = {1, 1, 1, 1, 1};
-   int *to_matrix[5];
+   double to_weights[5] = {1, 1, 1, 1, 1};
+   double *to_matrix[5];
    for (int i = 0; i < 5; i++)
    {
       to_matrix[i] = to_edges[i];
    }
  
-   Network to_network((int**)to_matrix, to_weights, 5);
+   Network to_network((double**)to_matrix, to_weights, 5);
    
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 1;
@@ -266,22 +266,22 @@ TEST(LineNetwork, SameNodeInvalidMapping)
 
 TEST(LineNetwork, InvalidLinks)
 {
-   int from_edges[2][2] = 
+   double from_edges[2][2] = 
    {
       {0, 5},
       {5, 0}
    };
-   int from_weights[2] = {1, 1};
-   int *from_matrix[2];
+   double from_weights[2] = {1, 1};
+   double *from_matrix[2];
    for (int i = 0; i < 2; i++)
    {
       from_matrix[i] = from_edges[i];
    }
  
-   Network from_network((int**)from_matrix, from_weights, 2);
+   Network from_network((double**)from_matrix, from_weights, 2);
 
 
-   int to_edges[5][5] =
+   double to_edges[5][5] =
    { 
      {0, 2, -1, -1, -1},
      {2, 0, 2, -1, -1},
@@ -290,14 +290,14 @@ TEST(LineNetwork, InvalidLinks)
      {-1, -1, -1, 2, 0}
    };
 
-   int to_weights[5] = {1, 1, 1, 1, 1};
-   int *to_matrix[5];
+   double to_weights[5] = {1, 1, 1, 1, 1};
+   double *to_matrix[5];
    for (int i = 0; i < 5; i++)
    {
       to_matrix[i] = to_edges[i];
    }
  
-   Network to_network((int**)to_matrix, to_weights, 5);
+   Network to_network((double**)to_matrix, to_weights, 5);
    
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 0;
@@ -308,7 +308,7 @@ TEST(LineNetwork, InvalidLinks)
 
 TEST(LineNetwork, InvalidMapping)
 {
-   int edges[5][5] =
+   double edges[5][5] =
    { 
      {0, 2, -1, -1, -1},
      {2, 0, 2, -1, -1},
@@ -317,15 +317,15 @@ TEST(LineNetwork, InvalidMapping)
      {-1, -1, -1, 2, 0}
    };
 
-   int *matrix[5];
+   double *matrix[5];
    for (int i = 0; i < 5; i++)
    {
       matrix[i] = edges[i];
    }
 
-   int nodes[] = {0,0,0,0,0};
-   Network from_network((int**)matrix, nodes, 5);
-   Network to_network((int**)matrix, nodes, 5);
+   double nodes[] = {0,0,0,0,0};
+   Network from_network((double**)matrix, nodes, 5);
+   Network to_network((double**)matrix, nodes, 5);
    NetworkMapping mapping(&from_network, &to_network);
    mapping.m_mapping[0] = 1;
    mapping.m_mapping[1] = 2;
